+++ protocols/datalink/frame_relay.h
@@ -17,7 +17,7 @@
 #define FRELAY_CR(pkt)          ((pkt[1] & 0x20) >> 5)  /* command/response bit */
 #define FRELAY_FECN(pkt)        ((pkt[0] & 0x02) >> 1)
 #define FRELAY_BECN(pkt)        (pkt[0] & 0x01)
-#define FRELAY_PROTO(pkt, len)  (((pkt[len] << 8) | pkt[len + 1]))
+#define FRELAY_PROTO(pkt, len)  ((uint16_t)(pkt[len] << 8) | (uint16_t)pkt[len + 1])
 
 #define NLPID_THRESHOLD         0xfe
