+++ protocols/transport/udp.h
@@ -10,10 +10,10 @@
 #define UDP_LENGTH_LABEL        "Length"
 #define UDP_CHECKSUM_LABEL      "Checksum"
 
-#define UDP_SRC_PORT(pkt)       ((pkt[0] << 8) | pkt[1])
-#define UDP_DEST_PORT(pkt)      ((pkt[2] << 8) | pkt[3])
-#define UDP_LENGTH(pkt)         ((pkt[4] << 8) | pkt[5])
-#define UDP_CHECKSUM(pkt)       ((pkt[6] << 8) | pkt[7])
+#define UDP_SRC_PORT(pkt)       ((uint16_t)(pkt[0] << 8) | (uint16_t)pkt[1])
+#define UDP_DEST_PORT(pkt)      ((uint16_t)(pkt[2] << 8) | (uint16_t)pkt[3])
+#define UDP_LENGTH(pkt)         ((uint16_t)(pkt[4] << 8) | (uint16_t)pkt[5])
+#define UDP_CHECKSUM(pkt)       ((uint16_t)(pkt[6] << 8) | (uint16_t)pkt[7])
 
 #define UDP_LEN                 8
